requirements.txt
praat-parselmouth
matplotlib
Numpy
openai-whisper